Restructure main(), add plcaeholder loops
listing imports is separated out into its own function, and an operation
function pointer is created, to make it easier to toggle between
different actions with options.

Operations on the source files is separated out of the loop where we lex
module files; comments are added to explain why this is necessary.

Various TODO comments are added to make it clear what is unfinished.