Reorganize source structure into flat module
Having a cmd/hareimports path with two three files in it gives me Java
nightmares.

Additionally, I re-added the list() function back to main.ha, as it is
simple and will not rely on the helper functions that the operations in
editing.ha will.

I also exported the operation functions.