Use arguments as paths to operate on, not module path
An alternate module path can ot be provided with the -m flag.

This interface will be more cohesive to using hareimports to
automatically format a file as it is being edited in a text editor, an
intended use case of hareimports.

A current known limitation is that the file paths provided must be
canonical (no symlinks) and absolute. It is planned to change this so
that the paths are checked to see if they resolve to the same file,
instead.

The -a option is added to tersely perform the previous behavior, of
operating on all source files in the module.